refactor(server): move ai types
