+++ packages/amplication-server/src/core/ai/ai.types.ts
@@ -1,0 +1,17 @@
+import { Entity, EntityField, Resource } from "../../prisma";
+
+type EntityFieldPartial = Pick<
+  EntityField,
+  "name" | "displayName" | "dataType" | "properties"
+>;
+
+export interface EntityPartial
+  extends Pick<Entity, "id" | "name" | "displayName"> {
+  versions: {
+    fields: EntityFieldPartial[];
+  }[];
+}
+
+export interface ResourcePartial extends Pick<Resource, "id" | "name"> {
+  entities: EntityPartial[];
+}
