+++ packages/amplication-server/src/core/ai/prompt-manager.types.ts
@@ -1,20 +1,3 @@
-import { Entity, EntityField, Resource } from "../../prisma";
-
-type EntityFieldPartial = Pick<
-  EntityField,
-  "name" | "displayName" | "dataType" | "properties"
->;
-
-interface EntityPartial extends Pick<Entity, "id" | "name" | "displayName"> {
-  versions: {
-    fields: EntityFieldPartial[];
-  }[];
-}
-
-export interface ResourcePartial extends Pick<Resource, "id" | "name"> {
-  entities: EntityPartial[];
-}
-
 interface DataModel {
   name: string;
   fields: {
